Fix user panel role display in admin panel.
The user roles in the admin panel usesd to be added using the p tag. However, this means that the align-middle class of the table doesn't actually center the text due to the margin-bottom applied to the p tags by default. I have changed the p tags to divs and now the text is property centered.